Split out USB from main convert source files to space indent
