+++ src/cdcacm.h
@@ -1,0 +1,2 @@
+void usb_init(void);
+void usb_poll(void);
